breaks out useLoadedPlan
